% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_SN.R
\name{compile_SN}
\alias{compile_SN}
\title{Post-process Semantic Norming Data for KS Dissertation
Post-process Semantic Norming from PsychoPy for Exp 2}
\usage{
compile_SN(filename)
}
\arguments{
\item{filename}{The Psychopy Output File. Should be a string (with '') with a .csv extension.}
}
\value{
A cleaned dataframe with individual trials for a given participant.
}
\description{
Post-process Semantic Norming Data for KS Dissertation
Post-process Semantic Norming from PsychoPy for Exp 2
}
\examples{
postprocess_FUNK('FK_Pilot2_SN2_testrun_2018_Nov_18_1023.csv')
This function was originally designed to compile semantic norming data for KS Dissertation.
Provides an "Expectancy Column" which is 1=matches intended semantic relation or 0=does NOT match intended semantic relation
}
