% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimhead.R
\name{dimhead}
\alias{dimhead}
\title{DIM HEAD}
\usage{
dimhead(df)
}
\arguments{
\item{df}{dataframe that you wish to apply this function to. Can also call a vector/matrix with data.frame(df)}
}
\value{
returns the dataframe DIMensions rows then columns & HEAD output first 6 rows
}
\description{
Print the output of all of these functions at once (dim and head in the same function).
}
\examples{
dimhead(CogLin_Data)
dimhead(data.frame(CogLin_Data))
}
