% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_values.R
\name{graph_values}
\alias{graph_values}
\title{Compute values for graphing}
\usage{
graph_values(df, sort_by, dep_var)
}
\arguments{
\item{df}{Dataframe. What data you want to use. Dataframe should be formatted so that each row has a dependent variable marked with the condition that you wish to sortby}

\item{sort_by}{String or list. What conditions you want your data summarized by; e.g. average by SNR etc.; can be a character "quotes" or a list c("a","b")}

\item{dep_var}{Dependent variable, column name that exists in dataframe}
}
\value{
datagrame that lists the N/sample size, mean, std deviation, std error
}
\description{
Takes dataframe and calculates descriptive statistics for a given dependent variable that is formatted for KS graphing with ggplot2
}
\examples{
graph_values(CogLin_Data,"Noise_Condition",Percent)

}
