% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hll_help.R
\name{hll_help}
\alias{hll_help}
\title{Hearing & Language Lab Help Tool
Get help on basic information in the Hearing & Language Lab}
\usage{
hll_help(string)
}
\arguments{
\item{string}{A query in string/character format such as "Email" or "Apple". Use hll_help("help") for more information. Input is NOT case sensitive}
}
\value{
Information associated with the query
}
\description{
Hearing & Language Lab Help Tool
Get help on basic information in the Hearing & Language Lab
}
\examples{
hll_help("help") will return information and provide all possible prompts
}
