% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Data}
\usage{
load_data(path)
}
\arguments{
\item{path}{the file path where all of your files are stored}
}
\value{
a data frame with all files concatenated
}
\description{
Imports multiple .csv files into R from a single, specified directory.
Files should all be .csv format and have the same columns and names.
Modified to include source file as a column
}
