% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_CogLAW.R
\name{postprocess_CogLAW}
\alias{postprocess_CogLAW}
\title{Postprocess CogLAW data - FORCED CHOICE}
\usage{
postprocess_CogLAW(filename, attribute_type, age_group)
}
\arguments{
\item{filename}{The Psychopy Output File. Should be a string (with '') with a .csv extension.}

\item{attribute_type}{A string that indicates whether the attribute condition was Familiar or Novel}

\item{age_group}{A string that indicates whether the participant is in the Younger or Older age group}
}
\value{
a cleaned dataframe with individual trials for a given participant. The overall accuracy for the subject in percent, accuracy for attributes vs. labels, & the accuracy by each target word. Attribute_Type and Age_Group columns will be added based on input parameters.
Created by KS 3/2018 for CS Undergraduate Thesis - CogLAW: Cognitive Learning Attributes & Words
Rev 4/2018 to include attribute and age group inputs
}
\description{
Postprocess CogLAW data for CS Thesis (3 argument version)
}
\examples{
postprocess_CogLAW('CogLAW_002_Order1A.csv',"Familiar","Young")

}
