% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_CogLAW_1input.R
\name{postprocess_CogLAW_1input}
\alias{postprocess_CogLAW_1input}
\title{Postprocess CogLAW data - FORCED CHOICE & 1 INPUT ONLY
Written by KS March 2018 for CS Thesis CogLAW - Cognitive Learning Attributes & Words}
\usage{
postprocess_CogLAW_1input(filename)
}
\arguments{
\item{filename}{The Psychopy Output File. Should be a string (with '') with a .csv extension.}
}
\value{
a cleaned dataframe with individual trials for a given participant. The overall accuracy for the subject in percent, accuracy for attributes vs. labels, & the accuracy by each target word
}
\description{
Postprocess CogLAW data - FORCED CHOICE & 1 INPUT ONLY
Written by KS March 2018 for CS Thesis CogLAW - Cognitive Learning Attributes & Words
}
\examples{
postprocess_CogLAW('CogLAW_002_Order1A.csv')
Note that this does not allow you to add attribute and age group which is not automated in the PsychoPy file

}
