% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_Context.R
\name{postprocess_Context}
\alias{postprocess_Context}
\title{Post-process CONTEXT data (KS QRP)}
\usage{
postprocess_Context(filename, noise_condition)
}
\arguments{
\item{filename}{The Psychopy Output File. Should be a string (with '') with a .csv extension.}

\item{noise_condition}{A string that indicates whether the listening condition was in noise or in quiet.}
}
\value{
A cleaned dataframe with individual trials for a given participant. The overall accuracy for the subject in percent by semantic expectancy condition (i.e.: Congruent (sem), Conflicitng (nonsem), Neutral). Can also be used for CONTEXT RHYME DATA
}
\description{
Post-process CONTEXT data (KS QRP)
}
\examples{
postprocess_Context('Context07_12SNR_testrun_2016_Jul_21_1600.csv',"noise")


}
