% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_EngDec.R
\name{postprocess_EngDec}
\alias{postprocess_EngDec}
\title{postprocess_EngDec}
\usage{
postprocess_EngDec(filename, order, group)
}
\arguments{
\item{filename}{The Psychopy Output File. Should be a string (with '') with a .csv extension. Must be written with correct participant label conventions (e.g. DegPhon_Subject #_ED_L#)}

\item{order}{The order that was assigned to the participant. This must be numeric.}

\item{group}{A string that indicates a group distinction (for example, group can be age or whether the participant is a child or adult)}
}
\value{
cleaned data for an individual participant to be added to larger data frame and accuracy based on the individual list.
}
\description{
For English Decision/Novel Label Sorting in fʌn.nɔlɛdʒi Exp. 1.1 (KS Dissertation) and DegPhon (CS Capstone)
}
\examples{
postprocess_EngDec('DegPhon_001_ED_L1_english_decision_2019_Feb_07_1038.csv')
postprocess_EngDec(file.choose()) # Use this for ease to pick out the file you want to use
}
