% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_FUNK.R
\name{postprocess_FUNK}
\alias{postprocess_FUNK}
\title{Post-process fʌn.nɔlɛdʒi (KS Dissertation)}
\usage{
postprocess_FUNK(filename)
}
\arguments{
\item{filename}{The Psychopy Output File. Should be a string (with '') with a .csv extension.}
}
\value{
A cleaned dataframe with individual trials for a given participant.
}
\description{
Post-process endpoint LWL accuracy from PsychoPy for fʌn.nɔlɛdʒi Exp. 1.2 (KS Dissertation)
}
\examples{
postprocess_FUNK('FK_Pilot2_L1_testrun_2018_Nov_18_1009.csv')
This function was originally designed for pilot data of Exp. 1.2 of KS Dissertation
}
