% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_LS.R
\name{postprocess_LS}
\alias{postprocess_LS}
\title{Post-process LEXICAL SEMANTIC (R56)}
\usage{
postprocess_LS(filename, AV_Condition)
}
\arguments{
\item{filename}{The Psychopy Output File. Should be a string (with '') with a .csv extension.}

\item{AV_Condition}{A string that indicates whether the AV condition was AV (audio + visual) or AO (audio only)}
}
\value{
A cleaned dataframe with individual trials for a given participant. AV_Condition will be added based on input parameters.
}
\description{
Post-process behavioral data for Lexical semantic task.
}
\examples{
postprocess_LS('LS001_A2_testrun_2018_Feb_17_1411.csv','AV')
}
