% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_LWL_E1.R
\name{postprocess_LWL_E1}
\alias{postprocess_LWL_E1}
\title{Postprocess fʌn.nɔlɛdʒi - EXP 1 (KS Dissertation)}
\usage{
postprocess_LWL_E1(filename, group)
}
\arguments{
\item{filename}{The Psychopy Output File. Should be a string (with '') with a .csv extension.}
}
\value{
A cleaned dataframe with individual trials for a given participant.
}
\description{
Post-process endpoint LWL accuracy from PsychoPy for fʌn.nɔlɛdʒi Exp. 1.2 (KS Dissertation)
Very similar to postprocess_FUNK() except there is a column for Trial number.
}
\examples{
postprocess_LWL_E1(file.choose(),"NH")
}
