% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_LWL.R
\name{restructure_LWL}
\alias{restructure_LWL}
\title{Restructure Looking While Listening (LWL) Data}
\usage{
restructure_LWL(dataframe)
}
\arguments{
\item{dataframe}{The dataframe (LWL output) that you want to modify. This MUST have the following columns: Subject, Condition (phonological, semantic, control), Trial, Timebin, Fixation (0, 0.5 or 1), target, competitor, distractor.
Your areas of interest (target, competitor, distractor) should be columns filled with 0s. These will populate via this function.}
}
\value{
The same dataframes with the columns target, competitor, distractor filled based on specific parameters.
}
\description{
Restructure DataWiz Output for Looking While Listening for Lexical Semantic (R56). iCoder is made to measure looking/fixations across two targets. This allows us to look at fixations across three possible targets.
}
\examples{
restructured_dataframe <- restructure_LWL(dataframe_name);
Be sure to create your area of interest columns that need to be labelled: target, competitor, distractor (case-sensitive).
It should like like this: dataframe_name$target <- 0
Frankly, the reason this is a function is because the original R script is 200 lines. This is a VERY context specific function.
}
