% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimheadtail.R
\name{dimheadtail}
\alias{dimheadtail}
\title{DIM HEAD TAIL}
\usage{
dimheadtail(df)
}
\arguments{
\item{df}{dataframe that you wish to apply this function to. Can also call a vector/matrix with data.frame(df)}
}
\value{
returns the dataframe DIMensions rows then columns; HEAD output first 6 rows; TAIL output last 6 rows
}
\description{
Print the output of all of these functions at once. Redundant but useful.
}
\examples{
dimheadtail(CogLin_Data)
dimheadtail(data.frame(CogLin_Data))

}
